Added TestBase class and refactoring old tests
