+++ addressbook-web-tests/src/test/java/com/lesson/addressbook/GroupDeletionTests.java
@@ -1,0 +1,15 @@
+package com.lesson.addressbook;
+
+import org.testng.annotations.Test;
+
+public class GroupDeletionTests extends TestBase{
+
+    @Test
+    public void testGroupDeletion() {
+        gotoGroupPage();
+        selectGroup();
+        deleteGroup();
+        returnGroupPage();
+    }
+
+}
